+++ contrib/packaging/startos/startos/actions/index.ts
@@ -1,5 +1,6 @@
 import { sdk } from '../sdk'
 import { caCertificate } from './caCertificate'
+import { mcpHostnames } from './mcpHostnames'
 import { mcpToken } from './mcpToken'
 import { network } from './network'
 
@@ -7,3 +8,4 @@ export const actions = sdk.Actions.of()
   .addAction(network)
   .addAction(caCertificate)
   .addAction(mcpToken)
+  .addAction(mcpHostnames)